[MISC] Various minor improvements (#1002)
* More robust successive init / destroy of Genesis (incl. viewer)
* Device management in RL examples has been improved to support Apple Metal
* Enable self-collision by default for Mujoco consistency
* Restore the ability to get entity com-based linear link's velocity
* Improve collision handling for inverse kinematics
* Warn about out-of-bound initial robot qpos
* Fix broken link reference in joint

+++ genesis/vis/rasterizer.py
@@ -123,13 +123,14 @@ class Rasterizer(RBC):
         self._context.jit.update_buffer(buffer_updates)
 
     def destroy(self):
-        if self._offscreen:
+        if self._offscreen and self._renderer is not None:
             self._renderer._platform.make_current()
             for target in self._camera_targets:
                 self._camera_targets[target].delete()
             self._renderer.delete()
             del self._renderer
             gc.collect()
+            self._renderer = None
 
     @property
     def viewer(self):
